+++ packages/hadron-build/src/lib/target.spec.ts
@@ -1,10 +1,9 @@
-'use strict';
-const path = require('path');
-const { expect } = require('chai');
-const Target = require('../src/lib/target').default;
+import path from 'path';
+import { expect } from 'chai';
+import Target from './target';
 
 describe('target', function () {
-  let env;
+  let env: NodeJS.ProcessEnv;
 
   beforeEach(function () {
     env = { ...process.env };
@@ -15,7 +14,7 @@ describe('target', function () {
   });
 
   it('should have autoupdate endpoint resolved from the package.json config', function () {
-    const target = new Target(path.join(__dirname, 'fixtures', 'hadron-app'));
+    const target = new Target(path.join(__dirname, '..', '..', 'test', 'fixtures', 'hadron-app'));
 
     expect(target).to.have.property(
       'autoUpdateBaseUrl',
@@ -24,7 +23,7 @@ describe('target', function () {
   });
 
   it('defaults to package.json distribution config options', function () {
-    const target = new Target(path.join(__dirname, 'fixtures', 'hadron-app'));
+    const target = new Target(path.join(__dirname, '..', '..', 'test', 'fixtures', 'hadron-app'));
 
     expect(target).to.have.property('distribution', 'compass');
     expect(target).to.have.property('name', 'compass');
@@ -47,7 +46,7 @@ describe('target', function () {
       HADRON_APP_VERSION: '1.2.3',
     });
 
-    const target = new Target(path.join(__dirname, 'fixtures', 'hadron-app'));
+    const target = new Target(path.join(__dirname, '..', '..', 'test', 'fixtures', 'hadron-app'));
 
     expect(target).to.have.property('distribution', 'compass-isolated');
     expect(target).to.have.property('name', 'compass-compass');
