--- packages/hadron-build/test/helpers.js
@@ -1,9 +1,0 @@
-'use strict';
-const Target = require('../src/lib/target').default;
-const path = require('path');
-
-exports.getConfig = (argv) => {
-  const src = path.join(__dirname, 'fixtures', 'hadron-app');
-  process.env.HADRON_DISTRIBUTION = 'compass';
-  return new Target(src, argv);
-};
